streamlit-app/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ctsr-streamlit"
version = "0.1.0"
description = "Clinical Trial Systems Register - Streamlit Frontend"
requires-python = ">=3.11"
authors = [
    {name = "CTSR Team"}
]

dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "plotly>=5.17.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pylint>=2.17.0",
]
